Remove legacy exec service and move new service out of wip
