+++ api/server/routes.go
@@ -11,7 +11,6 @@ import (
 	"github.com/rs/zerolog/log"
 	middleware2 "github.com/unweave/unweave/api/middleware"
 	"github.com/unweave/unweave/api/router"
-	"github.com/unweave/unweave/api/types"
 	"github.com/unweave/unweave/db"
 	"github.com/unweave/unweave/runtime"
 )
@@ -19,34 +18,6 @@ import (
 type Config struct {
 	APIPort string    `json:"port" env:"UNWEAVE_API_PORT"`
 	DB      db.Config `json:"db"`
-}
-
-func HandleRestart(ctx context.Context, rti runtime.Initializer) error {
-	// Re-watch all sessions
-	sessions, err := db.Q.ExecGetAllActive(ctx)
-	if err != nil {
-		return err
-	}
-
-	log.Ctx(ctx).Info().Msgf("🔄 Restarting watching %d sessions", len(sessions))
-
-	for _, s := range sessions {
-		sess := s
-		go func() {
-			c := context.Background()
-			c = log.With().
-				Str(types.UserIDCtxKey, sess.CreatedBy).
-				Str(types.ProjectIDCtxKey, sess.ProjectID).
-				Str(types.ExecIDCtxKey, sess.ID).
-				Logger().WithContext(c)
-
-			srv := NewCtxService(rti, "", sess.CreatedBy)
-			if e := srv.Exec.Watch(c, sess.ID); e != nil {
-				log.Ctx(ctx).Error().Err(e).Msgf("Failed to watch session")
-			}
-		}()
-	}
-	return nil
 }
 
 func API(cfg Config, rti runtime.Initializer, execRouter *router.ExecRouter) {
@@ -114,9 +85,6 @@ func API(cfg Config, rti runtime.Initializer, execRouter *router.ExecRouter) {
 
 	ctx := context.Background()
 	ctx = log.With().Logger().WithContext(ctx)
-	if err := HandleRestart(ctx, rti); err != nil {
-		panic(err)
-	}
 
 	log.Info().Msgf("🚀 API listening on %s", cfg.APIPort)
 	if err := http.ListenAndServe(":"+cfg.APIPort, r); err != nil {
