+++ main.go
@@ -11,8 +11,8 @@ import (
 	"github.com/unweave/unweave/api/server"
 	"github.com/unweave/unweave/db"
 	"github.com/unweave/unweave/providers/lambdalabs"
+	execsrv "github.com/unweave/unweave/services/exec"
 	"github.com/unweave/unweave/tools/gonfig"
-	execsrv "github.com/unweave/unweave/wip/services/exec"
 )
 
 func main() {
